Clear Sonar findings: S1659 and S1905 in sidecar
- SidecarReader: split four IndexInput declarations onto separate lines (S1659).
- SidecarBuilder line 61: drop redundant (long) casts; MIN_BUFFER_SIZE_MB is already
  long and Integer.MAX_VALUE auto-widens in Math.min(int, long) (S1905).
- SidecarReader lines 72, 92: drop redundant (long) casts on int operands of
  '* 8L' multiplications; binary numeric promotion already widens the int to
  long before the multiply, so overflow safety is preserved by the 8L literal
  alone (S1905).

+++ SearchSnapshotExtractor/src/main/java/org/opensearch/migrations/bulkload/lucene/sidecar/SidecarBuilder.java
@@ -58,7 +58,7 @@ public final class SidecarBuilder implements PostingsSink, AutoCloseable {
 
         // BufferSize.megabytes takes int MiB. Clamp to [MIN_BUFFER_SIZE_MB, Integer.MAX_VALUE].
         long mb = sortBufferBytes > 0 ? (sortBufferBytes >>> 20) : DEFAULT_SORT_BUFFER_MB;
-        mb = Math.max((long) OfflineSorter.MIN_BUFFER_SIZE_MB, Math.min((long) Integer.MAX_VALUE, mb));
+        mb = Math.max(OfflineSorter.MIN_BUFFER_SIZE_MB, Math.min(Integer.MAX_VALUE, mb));
         this.sortBufferSize = BufferSize.megabytes((int) mb);
 
         this.termsOut        = dir.createOutput(TERMS_FILE, IOContext.DEFAULT);
